dev-requirements.txt
homeassistant-stubs
flake8
mypy
